requirements.txt
# ================================
# Agentic AI Workflow Automation
# requirements.txt
# ================================

# --- Core Agent Framework ---
crewai==0.28.8
crewai-tools==0.1.6

# --- LLM (Groq - Free & Fast) ---
langchain-groq==0.1.3
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# --- Database (Supabase) ---
supabase==2.4.2

# --- UI (Streamlit) ---
streamlit==1.33.0

# --- Google APIs (Calendar + Gmail) ---
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0

# --- Slack ---
slack-sdk==3.27.1
requests==2.31.0

# --- Environment Variables ---
python-dotenv==1.0.1

# --- Utilities ---
pydantic==2.7.0
typing-extensions==4.11.0